Renamed folder to be more site-agnostic
